Cambio en lógica de pedidos y modificación de interfaz gráfica

+++ megazarl-presentacion/src/main/java/edu/student/itson/dissof/megazarl/presentacion/interfaces/IProductosVenta.java
@@ -9,7 +9,7 @@ public interface IProductosVenta {
 
     public abstract void setProductosBusqueda(List<Map<String, Object>> listaInformacionProductos);
 
-    public abstract void mostrarInformacionProducto(Integer idProducto);
+    public abstract void mostrarInformacionProducto(Long idProducto);
     
     public abstract void mostrarAvisoSinProductosDisponibles();
